Backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
django-filter>=23.0
hypothesis>=6.0
pytest-django>=4.7
Pillow>=10.0
python-decouple>=3.8
